Hisu

+++ src/main/java/group1/entity/SanPham.java
@@ -11,7 +11,7 @@ public class SanPham {
     public SanPham() {
     }//////
 
-    public SanPham( String tenSP,int soluong, float gia) {
+    public SanPham(String tenSP, int soluong, float gia) {
         this.maSP = maSP;
         this.tenSP = tenSP;
         this.gia = gia;
@@ -56,7 +56,7 @@ public class SanPham {
         this.anh = anh;
     }
 
-    public double getGia() {
+    public float getGia() {
         return gia;
     }
 
@@ -74,8 +74,7 @@ public class SanPham {
 
     @Override
     public String toString() {
-        return super.toString(); 
+        return super.toString();
     }
 
- 
 }
